Harden bounded step-log tails
Address review findings around bounded reads, cold fallback, and compaction publication safety.

Refs ENG-1836

+++ libs/api/logs/drizzle/0003_add_line_count_to_attempt_streams.sql
@@ -1,1 +1,1 @@
-ALTER TABLE "logs_attempt_streams" ADD COLUMN "line_count" integer;
+ALTER TABLE "logs_attempt_streams" ADD COLUMN "line_count" bigint;
